isolated view model logic, simplifyed server, created bot for single player mode
