+++ server/user.py
@@ -1,7 +1,6 @@
 class User:
-    def __init__(self, websocket, playing, pos):
+    def __init__(self, websocket, playing):
         self.websocket = websocket
         self.playing = playing
-        self.pos = pos
